rename sampleApp package from tech.dojo.pay.sdk to tech.dojo.pay.sdksample to cease conflicts between sampleApp and SDK
