+++ sample/build.gradle
@@ -7,7 +7,7 @@ android {
     compileSdk 32
 
     defaultConfig {
-        applicationId "tech.dojo.pay.sdk"
+        applicationId "tech.dojo.pay.sdksample"
         minSdk 21
         targetSdk 32
         versionCode 1
