+++ sample/src/main/java/tech/dojo/pay/sdksample/CardPaymentActivity.kt
@@ -1,5 +1,6 @@
-package tech.dojo.pay.sdk
+package tech.dojo.pay.sdksample
 
+import tech.dojo.pay.sdk.DojoSdk
 import tech.dojo.pay.sdk.card.entities.DojoCardPaymentPayload
 
 class CardPaymentActivity : CardPaymentBaseActivity() {
