+++ sample/src/main/java/tech/dojo/pay/sdksample/Cards.kt
@@ -1,4 +1,4 @@
-package tech.dojo.pay.sdk
+package tech.dojo.pay.sdksample
 
 import tech.dojo.pay.sdk.card.entities.DojoCardDetails
