+++ sample/src/main/java/tech/dojo/pay/sdksample/SampleActivity.kt
@@ -1,9 +1,10 @@
-package tech.dojo.pay.sdk
+package tech.dojo.pay.sdksample
 
 import android.content.Intent
 import android.os.Bundle
 import android.view.View
 import androidx.appcompat.app.AppCompatActivity
+import tech.dojo.pay.sdksample.R
 
 class SampleActivity : AppCompatActivity() {
