+++ sample/src/main/java/tech/dojo/pay/sdksample/token/TokenApi.kt
@@ -1,4 +1,4 @@
-package tech.dojo.pay.sdk.token
+package tech.dojo.pay.sdksample.token
 
 import retrofit2.http.Body
 import retrofit2.http.Headers
