+++ sample/src/main/java/tech/dojo/pay/sdksample/token/TokenResponse.kt
@@ -1,4 +1,4 @@
-package tech.dojo.pay.sdk.token
+package tech.dojo.pay.sdksample.token
 
 class TokenResponse(
     val id: String
