+++ sample/src/test/java/tech/dojo/pay/sdksample/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package tech.dojo.pay.sdk
+package tech.dojo.pay.sdksample
 
 import org.junit.Test
